cleanup of structure module
