+++ envinorma/io/docx.py
@@ -13,7 +13,7 @@ from bs4 import BeautifulSoup
 
 from envinorma.data import StructuredText
 from envinorma.data.text_elements import Cell, EnrichedString, Row, Table, TextElement, Title
-from envinorma.structure import build_structured_text
+from envinorma.from_legifrance import build_structured_text
 from envinorma.utils import random_string
