+++ tests/test_legifrance_to_am.py
@@ -1,11 +1,12 @@
 import json
 import random
 from datetime import date
+from typing import Union
 
 import pytest
 from leginorma import LegifranceSection, LegifranceText
 
-from envinorma.structure.am_structure_extraction import (
+from envinorma.from_legifrance.legifrance_to_am import (
     _BASE_LEGIFRANCE_URL,
     ArreteMinisteriel,
     ArticleStatus,
@@ -29,13 +30,7 @@ from envinorma.structure.am_structure_extraction import (
     remove_empty,
     remove_summaries,
     split_alineas_in_sections,
-    transform_arrete_ministeriel,
-)
-from envinorma.structure.texts_properties import (
-    _extract_section_inconsistencies,
-    count_articles_in_am,
-    count_sections,
-    count_tables,
+    legifrance_to_arrete_ministeriel,
 )
 from envinorma.utils import safely_replace
 
@@ -121,15 +116,13 @@ _SAMPLE_AM_NOR = ['DEVP1706393A', 'TREP1815737A', 'ATEP9870263A', 'DEVP1519168A'
 
 
 def _get_filename(nor: str) -> str:
-    return safely_replace(
-        __file__, 'tests/test_am_structure_extraction.py', f'test_data/AM/legifrance_texts/{nor}.json'
-    )
+    return safely_replace(__file__, 'tests/test_legifrance_to_am.py', f'test_data/AM/legifrance_texts/{nor}.json')
 
 
 @pytest.mark.filterwarnings('ignore')
 def test_no_fail_in_structure_extraction():
     for nor in _SAMPLE_AM_NOR:
-        transform_arrete_ministeriel(LegifranceText.from_dict(json.load(open(_get_filename(nor)))))
+        legifrance_to_arrete_ministeriel(LegifranceText.from_dict(json.load(open(_get_filename(nor)))))
 
 
 def test_weird_annexe_replacement():
@@ -202,7 +195,22 @@ def test_delete_or_merge_articles():
 
 def _get_am(filename: str) -> ArreteMinisteriel:
     raw_text = LegifranceText.from_dict(json.load(open(filename)))
-    return transform_arrete_ministeriel(raw_text, am_id=f'FAKE_ID_{filename}')
+    return legifrance_to_arrete_ministeriel(raw_text, am_id=f'FAKE_ID_{filename}')
+
+
+def count_sections(am: Union[ArreteMinisteriel, StructuredText]) -> int:
+    if not am.sections:
+        return 1
+    return sum([count_sections(section) for section in am.sections])
+
+
+def _count_structured_text_tables(text: StructuredText) -> int:
+    nb_tables = len([al for al in text.outer_alineas if al.table])
+    return nb_tables + sum([_count_structured_text_tables(section) for section in text.sections])
+
+
+def count_tables(am: ArreteMinisteriel) -> int:
+    return sum([_count_structured_text_tables(section) for section in am.sections])
 
 
 @pytest.mark.filterwarnings('ignore')
@@ -210,7 +218,6 @@ def test_structuration():
     am_1 = _get_am('test_data/AM/legifrance_texts/DEVP1706393A.json')
     assert count_sections(am_1) == 93
     assert count_tables(am_1) == 5
-    assert count_articles_in_am(am_1) == 14
     assert len(am_1.sections) == 14
 
 
@@ -219,7 +226,6 @@ def test_structuration_2():
     am_2 = _get_am('test_data/AM/legifrance_texts/TREP1835514A.json')
     assert count_sections(am_2) == 89
     assert count_tables(am_2) == 9
-    assert count_articles_in_am(am_2) == 60
     assert len(am_2.sections) == 6
 
 
@@ -227,25 +233,7 @@ def test_structuration_3():
     am_3 = _get_am('test_data/AM/legifrance_texts/fake_am.json')
     assert count_sections(am_3) == 101
     assert count_tables(am_3) == 1
-    assert count_articles_in_am(am_3) == 1
     assert len(am_3.sections) == 1
-
-
-def test_inconsistency_detection():
-    subsections = [
-        StructuredText(EnrichedString('1. Foo'), [], [], None),
-        StructuredText(EnrichedString('2. Bar'), [], [], None),
-        StructuredText(EnrichedString('3. Pi'), [], [], None),
-        StructuredText(EnrichedString('4. Pa'), [], [], None),
-    ]
-    section = StructuredText(EnrichedString(''), [], subsections, None)
-    inconsistencies = _extract_section_inconsistencies(section)
-    assert len(inconsistencies) == 0
-
-    subsections_err = [*subsections, StructuredText(EnrichedString('4. Pou'), [], [], None)]
-    section_err = StructuredText(EnrichedString(''), [], subsections_err, None)
-    inconsistencies_err = _extract_section_inconsistencies(section_err)
-    assert len(inconsistencies_err) == 1
 
 
 def test_structure_text():
@@ -259,16 +247,6 @@ def test_structure_text():
     assert text.sections[0].sections[0].title.text == 'A. pi'
     assert text.sections[0].sections[1].title.text == 'B. pa'
     assert text.sections[0].sections[2].title.text == 'C. po'
-
-
-# def test_structure_text_2():
-#     alineas = ['I. Foo', 'a) pi', 'hola', 'b) pa', 'quetal', 'c) po', 'II. Bar']
-#     text = _structure_text('', alineas)
-#     assert len(text.sections) == 2
-#     assert len(text.outer_alineas) == 0
-#     assert text.sections[0].title.text == 'I. Foo'
-#     assert text.sections[1].title.text == 'II. Bar'
-#     assert len(text.sections[0].sections) == 0
 
 
 def test_remove_summaries():
