+++ tests/test_structure_extraction.py
@@ -1,5 +1,5 @@
 from envinorma.data.text_elements import Table, Title
-from envinorma.structure import _build_enriched_alineas, _extract_highest_title_level, build_structured_text
+from envinorma.from_legifrance import _build_enriched_alineas, _extract_highest_title_level, build_structured_text
 
 
 def test_build_enriched_alineas():
